Create repository if missing and make deletion configurable.
This is a quick hack so tests can leave generated repository for later
use if generation of repository takes too long.

This is against idea that test-cases should be self-contained and
should not leave any garbage between test executions.

Also added code responsible for creation of initial commit. That
commit contains empty tree.

+++ src/test/scala/com/google/gimd/jgit/AbstractJGitTestCase.scala
@@ -28,18 +28,28 @@ abstract class AbstractJGitTestCase {
 
   protected var masterBranch: JGitBranch = null
 
+  /**
+   * Option determining weather repository created by this test-case will be deleted at the end
+   * of test-case execution.
+   */
+  protected val deleteRepository = true
+
   @Before protected def createRepository {
     val file = new File(repositoryPath)
     val repository = new Repository(file)
-    if (file.exists)
-      throw new IOException("Repository already exists: " + file)
-    file.mkdirs()
-    repository.create()
     masterBranch = JGitBranch(repository, masterRef)
+    if (!file.exists) {
+      file.mkdirs()
+      repository.create()
+      val treeId = addFiles(Nil)
+      val commitId = createCommit("Initial commit", treeId)
+      moveMaster(commitId)
+    }
   }
 
   @After protected def removeRepository {
-    org.apache.commons.io.FileUtils.deleteDirectory(new File(repositoryPath))
+    if (deleteRepository)
+      org.apache.commons.io.FileUtils.deleteDirectory(new File(repositoryPath))
   }
 
   protected def writeTextContent(text: String): ObjectId = {
